Base Program
This is the base program. It only contains a single window and attempts
to create a few folders on your system when it's launched. It can be
closed by clicking the `Kröw` menu button and then clicking the only
option: `close`.

The window has no decoration and is practically blank. There is nothing
to populate the `TableView`s that have been put on either side of the
window however the tables have been set up to display data as soon as it
is added with a simple function call.

There may or may not be a few errors in the loading methods that read
`Construct` objects from the filesystem at the moment; these features
will be setup in the following commit.

As a final note, there may be a few mistakes or "minor inconveniences"
in the code here and there. One way or another, if I learn about them, I
will fix them.

~Zeale