Themed the Program
The main change in this commit is the introduction of a custom CSS Theme
for JavaFX. It MIGHT use some things that are undocumented in the
`JavaFX CSS Reference` online, but I made sure to leave (semi-)helpful
comments everywhere. (And if I didn't, I probably will soon.)

Anyways, I plan to add more functionality in the next commit, since the
program is literally just a window that does nothing at the moment. That
however, may change if I find something else more urgent or more fun.

Note: On my Windows computer (and in the `SceneBuilder` application that
I use to generate FXML) the very bottom part of the window's
`BorderPane` has a white border around it, which I can't seem to remove.
I don't know what object this white border belongs to, but it seems to
be outlining the `ScrollPane` that's placed down there. The border also
seems to be on top of the `ScrollPane`, as it shows even when I give the
`ScrollPane` a custom border (which should show on top of this mystery
border). I'll mark this as an issue soon if I can't find a fix.

Note: On a less important note, the `TableView` object in the right side
of the window has an extra column of cells (which I've managed to make
very thin) whos rows get highlighted when any of the corresponding,
actual cells get highlighted. This is better explained when seen, but
the program currently has no way of populating the `TableView` as the
lists which control their contents are empty. This doesn't look too bad
and this is - as I said before - a minor issue, but it is noticeable.
Once again, I should add functionality to this program and allow the
user to actually do things that will populate these lists and more in
the next commit, but for now, I might add this little bug as a minor
issue. (One more thing: I googled this problem and found a few results
that didn't work for me. The "mystery border" issue, however, pulled up
nothing on google.)