Construct Creation
Constructs can now be created via the `CreateConstructWindow`. (The
`Marks...` button in that window does not work.)

To get to the `CreateConstructWindow` the user must click the `New` menu
at the top of the home window and then click `• Construct` . This will
open the `CreateConstructWindow`. To leave the `CreateConstructWindow`
the user must click its `Options` menu and then hit the only menu item:
`Back`.

Constructs can be created and will be saved on program exit. The program
can be safely closed via the Home window's `Kröw` menu's `Close` button.

Notes:
* Constructs can not be deleted.
* The mystery border issue still exists (I will officially assign an
issue for it tomorrow)
* I SLIGHTLY modified the `DefaultTheme.css` file to remove the bottom
border of the `.side-tables` since the `.bottom-pane` has a top border
which already shows. (Keeping both borders puts the borders side by side
showing a border SINGLE border which is twice as thick as each of the
individuals. The borders combine, basically.)
* The `Law` and `Construct` class were quite heavily edited (while still
maintaining the same API) to account for Serialization AND easy display
via a `TableView`. To display information in a `TableView` easily, I
made all the information into `*Type*Property`s (Such as
`StringProperty` and `BooleanProperty`). Since these objects are not
Serializable, I implemented the `readObject` and `writeObject` methods
(that are called automatically by Java's `ObjectOutputStream`s or
`ObjectInputStream` as long as they have the exact same method
signature) and wrote each propery's value.
* I removed that little test code from the `main(String[] args)` method
of `Kröw.java` since there is no longer any need to test
`ObjectOutputStreams` (because the program uses them each run
automatically :D) and because it was test code and shouldn't have been
in the commit... :)
* I implemented the `stop() throws Exception` method in `Window.java`.
It now saves all `Constructs` on exit. :D
* Also in the `Window.java` class, there are now static methods to
retrieve the:
1. `currentController` - The current controller of the open window. This
is a `Window` object.
2. `currentScene` - This was really, already retievable through the
current stage. (`Window.getStage()`)
3. `previousController` - This is the controller of the `Scene` that was
last open. It is a `Window` object.
4. `previousScene` - This is the `Scene` that was open before the
currently open `Scene`.
Please note that **for any of the above variables to be correctly set,
changing the currently open Scene must be done with one of the static
methods in `Window`. (Currently one of either `setScene(...)` or
`setSceneToPreviousScene()`.)
* Last but not least, (well,  this might be the least tbh...), all
windows are draggable via the `MenuBar` at the top.