Construct Manager Window
Major updates.

For starters, the program now has an icon, (which can be changed in the
options tab of the Home window.), but apart from that, here's a list of
other changes:

* There is a Construct Manager window. The window displays all sorts of
information pertaining to constructs and allows you to edit the
information of a construct. The window shows a table out of the box.
When any construct in the table is clicked, the user is shown various
editing options and dialogs to change the properties of the already
existing construct. There is also a delete option which will delete the
construct being edited **without any warnings**. Use with caution.
* Pie charts. These are on the Construct manager window, as described
above. There are two different pie charts: one that shows the overall
gender of constructs and one that compares the amount of living
constructs with the amount of dead. The pie charts along with the table
are updated whenever a construct is added, removed, or edited.
* Constructs now have an `alive` property. This determines whether or
not they ~like potatoes~ are alive.
* The program now has an icon which looks like a little black crow. This
can be changed to a white bird from the options tab at the top of the
Home window.
* In the Home window, there is some color changing text that tells you
how many constructs you have. This gets updated as expected.
* Also in the home window, there is an image of a crow with a swirly
yellow and gray background. Try hovering over it and dragging it around.
* There is a "Debug mode" which can be activated by calling the program
with the **Program arguments** of either `-debug-mode` or `-debug`.
* Now on the aspect of internal operations and actual code:
* The construct class has its own CellValueFactory object. This is
what's used to display items in the table in the
`ConstructManagerWindow` (not the Home Window).
* Serializing a Construct will now save the Construct's "Version." This
is a long which basically denotes the version of the Serialized object
(Similar to the SerialVersionUID). During deserialization, the program
will read the Version of the object it is loading and load in properties
as they would've been loaded at that specific version. If the object
does not contain any version information, the program will attempt to
load it as it would have been loaded before this update of the program.
This ensures that this program will be backwards compatible. Every time
information stored in a serialized class is changed, (specifically the
`Construct` class as of now, since no other objects are createable by
the program), the Version is incremented so that the serialized object
is distinguishable from the previous versions of it.
* Some CSS changes have taken place. The `DefaultTheme.css` no longer
contains `center-pane`, `side-tables`, or `bottom-pane` classes and I
beleive a few other classes have been extracted as well. These classes
were only used by the `Home` window, so I created a CSS file
specifically for the `Home` window and moved the classes there. The
`DefaultTheme.css` file is now a boilerplate.
* If the program is run in a version of Java where a few deprecated
methods are used (specifically the methods in `HomeWindow.java` which
stop the `TableColumns` of the Home window's `TableView` from being
reordered) the program will be able to continue but will print some
information to the console informing the user of what has happened.
These deprecated methods are used in the Construct Manager window as
well.
* The `Kröw` class has a few static methods that return lists of
`Construct`s that satisfy certain conditions. This might change soon,
mostly because the actual `getConstructs()` method is NOT static and
belongs to the `INSTANCE` object which is static, final, and is created
when that class is loaded up. (There are a few reasons that the program
is designed like this, but there might be some better workarounds.) The
other big reason that this might change (or be removed) is that I can
vaguely recall a method taking in a `Predicate` object of some sort
which might provide the exact functionality needed. It would be very
cluttery to pepper the `Kröw` class with static methods that get certain
types of Constructs, Laws, Systems, and basically every other thing that
the program supports.
* Constructs now have a private static `List` that stores all the
`Construct`s that have been edited since the program has started. These
`Construct`s are saved when the program is closed. This was made so that
the program could keep track of which `Construct`s it actually needs to
save, instead of saving every single one when it comes time to close, as
that would be unnecesarilly laggy for programs managing a large amoung
of `Construct`s. The `Law` class has the same implementation.
* The "two suppressed fields" in the `HomeWindow.java` file were removed
as they weren't used in this commit. ._. I'll try to use them for what I
planned to at some point, so if I do decide to use them, I'll add them
back. (Obviously...)
* This isn't really a change but I thought that it should be
mentioned... The program has virtually no documentation, anywhere. I
will likely start documenting certain areas of the program sending
commits bit by bit. The main problem I have is working on decoration and
not sticking to the main program. (Specifically adding Laws and better
functionality towards the already configurable Construct objects.)