Added Support for Laws (And much more, as usual.)
Ok. For starters,
1. Text wrapping was enabled on most (if not all) of the `TextArea`s in
the program.
2. Next, the `Kröw` class no longer holds the `ObservableList`s of
`Construct`s or `Law`s or any other one of those objects in the program.
The `Kröw` class now has a `getDataManager()` method which will return a
`DataManager` object. This object contains plenty of useful methods
allowing for much better functionality.
3. Those `objects in the program` that I just mentioned above (like
`Construct`s and `Law`s and `System`s etc.) are all now extensions of
the `MindsetObject` class.
4. There is now a Law creation window. Laws can actually be made now.
(If you don't put in the date, it will take in the current time. Down to
the nearest *millisecond*...)
5. As I just said, **laws can actually be made now!!** You can store
their description, name, rules, and the date that they were created.
6. There is a System creation window as well! Systems can be made using
a name, description, and creation date! (Just like laws, if no Date is
given, the current time to the nearest millisecond is used. Pretty
fancy, right?)
7. The `ConstructManager.fxml`, `ConstructManagerWindow.java`,
`GenderPieChart.css`, and `LifePieChart.css` files have all been moved
around. These changes aren't too big and they don't affect the API (or
whatever you want to call the jumble of nonsense that's supposed to
allow for greater functionality), but I thought I'd mention them.
8. Pie Charts in the Construct Manager Window now have labels. That you
can see. I was thinking of making those labels show percentages instead
of the already fairly obvious legend, but I haven't really looked into a
way for that yet.
9. There is a **Law Manager** window just like the **Construct Manager**
window. Even though the last commit basically centered around adding the
*Manager Window* for `Construct`s, I thought it would be better if I
just went through and added a window quite similar to that but for laws.
The Law Manager window does not contain pie charts... Because there is
nothing to show in pie charts, BUT, it does allow you to correctly edit,
and delete `Law`s. What number am I on...? Oh yes, 10.
10. There was an empty block of code in the `DefaultTheme.css` file.....
Keyword, "was".
11. I'm working on an API called *Pages* that can be given any
`MindsetObject`, and display all its information neatly on a page,
regardless of the Objec'ts dynamic type. That's going to be somewhat
hard though... The `ObjectPage.fxml` and `NavigationPage.fxml` are both
part of this API. The `Pages.java` file is as well. They are not meant
to be used yet, **however**, if one *really* wanted to, they could
enable debug mode and in the Home page, click the Kröw tab...
12. Some methods were reorganized around the project.
13. `MindsetObject`s now have a few methods having to do with Files.
(`getFile()`, `getExtension()`, etc.) These are mostly just used by the
program for saving or deleting an object.
14. `Law`s have a `deleted` boolean.
15. `Law`s have a custom `CellValueFactory` class.
16. The `System` class now exists. You can create `System`s but you
can't really see them or do anything with them yet. You can't even tell
they exist...
17. The `DataManager class`. It does what it says, manages data. Methods
and fields pertaining to data should be found in a `DataManager` object.
(Currently, `Kröw` has one which can be accessed with
`Kröw.getDataManager()`.)
18. Many MANY things were moved from the `Kröw` class to the
`DataManager` class. Be informed.
19. The `Window` class was very gently changed . It now saves everything
that needs to be saved when the application exits and its `stop()`
method is called.

This commit was huge, and much bigger than it needed to be. I'm going to
try to do things like this in much tinier but more plentiful commits.
Next I'll work on documentation and maybe reorganize some classes.