Documented Window class and various areas
The Window class was fully documented as of this commit, and some basic
MindsetObject classes were given documentation. (For the class itself,
not each individual part of the class.)

The two general additions of this commit are the Window class's ful
documentation, and the MindsetObjects' class header documentation. The
rest of the MindsetObjects should already be documented.