Remade Popup API
The previous Popup API was far too abstracted/broad. Popups are meant to
be quick, almost single-statement/expression calls in the code, since
they will be made so often; (for most Nodes in a UI). With the
abstracted API, too many repetetive calls were required for each Popup
being made.