Fixed menu exiting bug
Fixed bug caused by the slide menu mouse exit event being called when a
popup was exited in the slide menu. JavaFX would pass a (sub)child of
the slide menu, as the picked node, in an event that was called for the
parent node (slide menu) being exited. The slide menu wasn't exited,
however, since one of its children were picked.