Modified SelectableCell hitbox
Made the entire SelectableCell respond to mouse clicks.

Clicking with the primary mouse button will reverse whether or not the
check box is selected.
Clicking the middle mouse button will always remove the checkbox's check
as well (the checkbox will be set to false), as long as the click isn't
actually on the check box itself; (only the cell was given a mouse event
listener; not the check box itself).