Added a refresh button for Schedule Module
This refresh button seems to work fine; it appears to accurately order
all of the items in the table.

Live refreshing is probably possible but, since table items teleport
around as soon as they're edited when live updating is enabled, they are
hard for the user's eyes to follow. When live updating is enabled, cells
no longer bind their checkboxes to items above them or anything like
that, which is good, but there seems to be a large amount of lag when
checking and unchecking some items when live updating is enabled.