Added Protection keys and Protection protection
The Protection API has been protected... or, made more secure. The
Domain class now only takes ProtectionKey objects, whose constructors
are private to the Protection class. This was implemented as a secure
way for the Protection class to communicate with the Domain class.

This limits the ability of non-malicious (that seems too mean of a word)
third-parties or clients to build on top of or try and change the way
that the API works, maybe by issuing Domains when calling methods
satisfy custom requirements, rather than the default set by the
Protection API. (That a class has to define who can use its domain with
annotations, and that a class is automatically allowed to use its own
domain, always.) Because of the limitations set by this little security
thing, it may be removed or changed in some way. Then again, people can
just instantiate this ProtectionKey with reflection... >:(