Added a loading/splash screen (No progress updates or texture)
There is an undecorated splash screen that shows loading progress. It
has a test image with a progress bar under it. The progress bar turns
into a button when the application is done loading and is ready to be
shown. The progress bar is not updated.

I made this mostly as a result of the speed of the computers at my
school. When I try to show my app to one of my classmates, it takes,
like, 2 whole fricking minutes to load, like I'm running my app on a
nokia processor or something.

This little splash screen thing is designed to show the user that the
app did actually open, and I added in a loading bar, because why not? So
now the user can also see how loading is progressing.

I may have it measured startup loading time and print it to the console.
Or maybe to the loading scene. It's literally like 3 lines of code
anyways. Probably two if I really want it to be. I'd love to compare my
3.6 GHz processor to my school's single core, 10MHz CPUs.