Dragging now works correctly
I may just store nodes in wrappers and have Windows keep a list of the
wrappers rather than the nodes themselves. The wrappers could be used to
manipulate effects and such, and effects applied pertaining to
"editMode" could have their initial values cached in the wrappers.
Something like that.

Next, a WindowBuilder window will close its children (the regular,
editable windows it owns) whenever it closes. Once a window is
"savable," and its state can be saved to the hard drive, I can make a
popup that asks if "you'd really like to close the window(s)?"