Compressed NumberPrompt input algorithm
So I compressed this but there's an error with it, particularly when I
try to scroll down from a number between 100 and 199. It decrements by
100, rather than by 10. This also happens for other first order ranges,
such as 1000 and 1999 or 10000 and 19999 etc, because of the way the
algorithm works.

Not to sure whether or not to count this as a bug.