ChatRoom (probably) works correctly
The remodel didn't work out EXACTLY like I wanted it to, but it's 1:46
AM and I don't feel like killing myself over reprogramming this. I'm
gonna make the command portion of the ChatRoom into its own little class
and then (probably) be done with abstraction from there. What I hoped
was to create an API where the ChatRoom class could fill out like two or
three abstract methods which would govern the functionality, but the
chat room just has too many features.

Once I know more I may be able to remodel this correctly. But for now,
this should work fine. I'm not too happy that the code isn't very much
less cluttered than what it was before the remodeling, but I'm more
worried about the time.

Anyways, the hover popups need to be reprogrammed so that they only
appear when your mouse stays still for some time and so that they
disappear when the mouse moves. Trying to click on an app and getting my
click event eaten by the popup is getting tedious. Also, the math module
needs to handle derivatives and integrals. I'm too lazy to go and type
out these features/updates in the Projects portion of Kröw's GitHub
repo, so...